pyproject.toml
[project]
name = "genai-ogiri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "google-genai>=1.5.0",
    "openai>=1.65.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.27.0",
    "pyyaml>=6.0.1",
]
